Updated to support Azure Functions V2
Updated to resolve error 'The function runtime is unable to start.' when trying to run PullEnvironments Function.

The host.json file was missing the required 'version' property. Have updated to V2 to include the missing version 2.0 property  and as per the V2 specifications have moved the application-level extension settings into "extensions"

Refer to
https://aka.ms/functions-hostjson